requirements.txt

v==1.1.0
Flask-Migrate==4.1.0
flask-marshmallow==1.3.0
marshmallow==4.0.0
